+++ tests/test_utils.py
@@ -26,3 +26,10 @@ class TestContextToCharStar(TestCase):
     def test_convert_to_char(self):
         for elem in ("", u'hi all'):
             self.assertEqual(six.b(elem), x2char_star(elem))
+
+    def test_convert_all(self):
+        self.assertEqual(b'true', x2char_star(True, convert_all=True))
+        self.assertEqual(b'false', x2char_star(False, convert_all=True))
+        self.assertEqual(b'1', x2char_star(1, convert_all=True))
+        self.assertEqual(b'3.14159', x2char_star(3.14159, convert_all=True))
+        self.assertEqual(b'100', x2char_star(100, convert_all=True))