+++ CHANGES.md
@@ -8,8 +8,10 @@ the one you find pypi, but its the one in development.
 =====
 * Add raise_on_fail to `pystemd.run` to raise a error when unit exit with
   non-zero status.
-* change on error when property probided to `dbus.call_method` is not the right
+* change on error when property provided to `dbus.call_method` is not the right
   type, still much work to do here.
+* added pystemd.daemon.notify as interface to sd_notify.
+* added pystemd.daemon.listen_fds as interface to sd_listen_fds.
 
 0.3.0
 =====